#1296 Fix test package structure.
