+++ foreign-key-mapping/src/test/java/com/iluwatar/foreignkeymapping/OrderTest.java
@@ -24,12 +24,5 @@
  */
 package com.iluwatar.foreignkeymapping;
 
-public interface AppDbSimulator {
-  Object find(int id, String table);
-
-  void insert(Object object, String table);
-
-  void update(Object object, String table);
-
-  void delete(int id, String table);
+public class OrderTest {
 }
